Upgrade Go to 1.23 (#6647)
* Update some for loops in response to the following change introduced
  in Go 1.22: each iteration of the loop creates new variables, to avoid
  accidental sharing bugs. As a result of this change, a few existing
  loop variables had become heap-allocated, so we updated the loop to
  avoid the extra allocation. Other loops that were affected by the
  change (compiling differently), but for which the loop variable was
  stack-allocated, were reviewed individually and when it made sense,
  the loop was refactored to avoid the extra variable copy. The `go test
  -gcflags=-d=loopvar=2` command was used to find affected loops.
* The antrea/codegen image was updated to use Go 1.23. No changes in
  generated files.
* golangci-lint was updated to the latest available version (v1.60.3)

    - govet reported some invalid usages of printf, which was addressed;
      this lead to some refactoring of the `antctl check` code
    - a new gosec rule (G115), which looks for potential integer
      overflows when converting between integer types, caused a lot of
      errors to be reported in the codebase. There were too many errors
      to be addressed as part of this PR, so we only handled a few of
      them for which the fix was straightforward. When the rule
      implementation is improved (so that it reports fewer false
      positives), we should review all errors and address them as needed.

Fixes #6644

+++ pkg/util/ip/ip.go
@@ -96,7 +96,7 @@ func diffFromCIDR(allowCIDR, exceptCIDR *net.IPNet) []*net.IPNet {
 	remainingCIDRs := make([]*net.IPNet, 0, exceptPrefix-allowPrefix)
 	for i := allowPrefix + 1; i <= exceptPrefix; i++ {
 		// Flip the (ipBitLen - i)th bit from LSB in exceptCIDR to get the IP which is not in exceptCIDR
-		ipOfNewCIDR := flipSingleBit(&exceptStartIP, uint8(bits-i))
+		ipOfNewCIDR := flipSingleBit(&exceptStartIP, bits-i)
 		newCIDRMask := net.CIDRMask(i, bits)
 		for j := range allowStartIP {
 			ipOfNewCIDR[j] = allowStartIP[j] | ipOfNewCIDR[j]
@@ -108,10 +108,10 @@ func diffFromCIDR(allowCIDR, exceptCIDR *net.IPNet) []*net.IPNet {
 	return remainingCIDRs
 }
 
-func flipSingleBit(ip *net.IP, bitIndex uint8) net.IP {
+func flipSingleBit(ip *net.IP, bitIndex int) net.IP {
 	newIP := make(net.IP, len(*ip))
 	copy(newIP, *ip)
-	byteIndex := uint8(len(newIP)) - (bitIndex / 8) - 1
+	byteIndex := len(newIP) - (bitIndex / 8) - 1
 	// XOR bit operation to flip
 	newIP[byteIndex] = newIP[byteIndex] ^ (1 << (bitIndex % 8))
 	return newIP
